@@ -11,16 +11,16 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.3.8",
-    "@angular/cdk": "^13.3.7",
-    "@angular/common": "~13.3.8",
-    "@angular/compiler": "~13.3.8",
-    "@angular/core": "~13.3.8",
-    "@angular/forms": "~13.3.8",
-    "@angular/material": "^13.3.7",
-    "@angular/platform-browser": "~13.3.8",
-    "@angular/platform-browser-dynamic": "~13.3.8",
-    "@angular/router": "~13.3.8",
+    "@angular/animations": "~14.2.6",
+    "@angular/cdk": "^14.2.5",
+    "@angular/common": "~14.2.6",
+    "@angular/compiler": "~14.2.6",
+    "@angular/core": "~14.2.6",
+    "@angular/forms": "~14.2.6",
+    "@angular/material": "^14.2.5",
+    "@angular/platform-browser": "~14.2.6",
+    "@angular/platform-browser-dynamic": "~14.2.6",
+    "@angular/router": "~14.2.6",
     "bootstrap": "^5.1.3",
     "popper.js": "^1.16.1",
     "rxjs": "~7.5.5",
@@ -30,7 +30,7 @@
   "devDependencies": {
     "@angular-devkit/build-angular": "~13.3.5",
     "@angular/cli": "~13.3.5",
-    "@angular/compiler-cli": "~13.3.8",
+    "@angular/compiler-cli": "~14.2.6",
     "@types/node": "^12.11.1",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.3",
