@@ -1,9 +1,9 @@
-import { FormControl } from '@angular/forms';
+import { UntypedFormControl } from '@angular/forms';
 import { Observable } from 'rxjs';
 
 export class CustomValidators {
 
-  static forbiddenUsernames(control: FormControl): {[key: string]: boolean} {
+  static forbiddenUsernames(control: UntypedFormControl): {[key: string]: boolean} {
     const forbiddenUsernamesList = ['admin', 'user', 'test'];
 
     if (forbiddenUsernamesList.includes(control.value)) {
@@ -12,7 +12,7 @@ export class CustomValidators {
     return null; // form control is valid
   }
 
-  static forbiddenEmailAsync(control: FormControl): Promise<any> | Observable<any> {
+  static forbiddenEmailAsync(control: UntypedFormControl): Promise<any> | Observable<any> {
     const promise = new Promise<any>((resolve, reject) => {
       setTimeout(() => {
         if (control.value === 'test@test.com') {
